Remove @chiselName and NoChiselNamePrefix (#2694)
