+++ macros/src/main/scala/chisel3/internal/naming/NamingAnnotations.scala
@@ -145,57 +145,6 @@ class NamingTransforms(val c: Context) {
     }
     """
   }
-
-  /** Applies naming transforms to vals in the annotated module or method.
-    *
-    * For methods, a hierarchical naming transform is used, where it will try to give objects names
-    * based on the call stack, assuming all functions on the stack are annotated as such and return
-    * a non-AnyVal object. Does not recurse into inner functions.
-    *
-    * For modules, this serves as the root of the call stack hierarchy for naming purposes. Methods
-    * will have chiselName annotations (non-recursively), but this does NOT affect inner classes.
-    *
-    * Basically rewrites all instances of:
-    * val name = expr
-    * to:
-    * val name = context.name(expr, name)
-    */
-  def chiselName(annottees: c.Tree*): c.Tree = {
-    var namedElts: Int = 0
-
-    val transformed = annottees.map(annottee =>
-      annottee match {
-        case q"$mods class $tpname[..$tparams] $ctorMods(...$paramss) extends { ..$earlydefns } with ..$parents { $self => ..$stats }" => {
-          val transformedStats = transformClassBody(stats)
-          namedElts += 1
-          q"$mods class $tpname[..$tparams] $ctorMods(...$paramss) extends { ..$earlydefns } with ..$parents { $self => ..$transformedStats }"
-        }
-        case q"$mods object $tname extends { ..$earlydefns } with ..$parents { $self => ..$body }" => {
-          annottee // Don't fail noisly when a companion object is passed in with the actual class def
-        }
-        // Currently disallow on traits, this won't work well with inheritance.
-        case q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" => {
-          val transformedExpr = transformHierarchicalMethod(expr)
-          namedElts += 1
-          q"$mods def $tname[..$tparams](...$paramss): $tpt = $transformedExpr"
-        }
-        case other =>
-          c.abort(
-            c.enclosingPosition,
-            s"@chiselName annotion may only be used on classes and methods, got ${showCode(other)}"
-          )
-      }
-    )
-
-    if (namedElts != 1) {
-      c.abort(
-        c.enclosingPosition,
-        s"@chiselName annotation did not match exactly one valid tree, got:\r\n${annottees.map(tree => showCode(tree)).mkString("\r\n\r\n")}"
-      )
-    }
-
-    q"..$transformed"
-  }
 }
 
 @compileTimeOnly("enable macro paradise to expand macro annotations")
@@ -205,12 +154,4 @@ class dump extends StaticAnnotation {
 @compileTimeOnly("enable macro paradise to expand macro annotations")
 class treedump extends StaticAnnotation {
   def macroTransform(annottees: Any*): Any = macro chisel3.internal.naming.DebugTransforms.treedump
-}
-@compileTimeOnly("enable macro paradise to expand macro annotations")
-@deprecated(
-  "Use chisel3.experimental.AffectsChiselPrefix instead. @chiselName will be removed in Chisel 3.6",
-  "Chisel 3.5"
-)
-class chiselName extends StaticAnnotation {
-  def macroTransform(annottees: Any*): Any = macro chisel3.internal.naming.NamingTransforms.chiselName
 }
